utils 작업
