+++ README.md
@@ -12,7 +12,7 @@
          ServiceController   - 서비스 관련 요청 처리
          CommandsController  - 커맨드 관련 오청 처리
          NagiosController    - 나기오스 상태, 재시작 요청 처리
-         StatisticController - 나기오스 상태
+         StatisticController - 나기오스 상태 개수, 로그
 
 # 미들웨어 변경 사항
     경로: app/Http/Kernel.php
