+++ app/Interfaces/NagiosInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Interfaces;
+
+/**
+ * Interface for API of nagios.
+ */
+interface NagiosInterface
+{
+    public function getCgiResult($sCommand);
+}
